+++ src/ops/mod.rs
@@ -1,4 +1,5 @@
 pub mod doc;
+pub mod file;
 pub mod md;
 pub mod patch;
 pub mod read;